[storage][diffs] diff manager files are renamed to diffs data source
