+++ storage/CMakeLists.txt
@@ -19,8 +19,8 @@ set(
   country_tree.hpp
   diff_scheme/apply_diff.cpp
   diff_scheme/apply_diff.hpp
-  diff_scheme/diff_manager.cpp
-  diff_scheme/diff_manager.hpp
+  diff_scheme/diffs_data_source.cpp
+  diff_scheme/diffs_data_source.hpp
   diff_scheme/diff_scheme_loader.cpp
   diff_scheme/diff_scheme_loader.hpp
   diff_scheme/diff_types.hpp
